architect: split cpp analyzer mutation hotspots
