+++ packages/core/src/treeSitter/runtime/analyzeCpp/relationDeclaredMethods.ts
@@ -1,0 +1,2 @@
+export { readCppDeclaredMethodNames } from './relationDeclaredMethodNames';
+export { readCppDeclaredMethodSymbols } from './relationDeclaredMethodSymbols';
