+++ packages/core/src/treeSitter/runtime/analyzeCpp/symbolDeclarationVariables.ts
@@ -1,0 +1,48 @@
+import type Parser from 'tree-sitter';
+import type { IAnalysisSymbol } from '@codegraphy-dev/plugin-api';
+import type { TreeWalkAction } from '../analyze/model';
+import { addNamedSymbol } from './symbolCreate';
+import { getDeclaratorNameNodes } from './symbolDeclaratorNames';
+import type { CppSymbolWalkState } from './symbolModel';
+import {
+  hasFunctionDeclarator,
+  isInsideClassLike,
+} from './symbolScope';
+
+export function handleCppDeclaration(
+  node: Parser.SyntaxNode,
+  filePath: string,
+  symbols: IAnalysisSymbol[],
+  state: CppSymbolWalkState,
+): TreeWalkAction<CppSymbolWalkState> | void {
+  if (state.currentFunctionSymbolId) {
+    addCppDeclarationSymbols(node, filePath, symbols, 'local');
+    return { skipChildren: true };
+  }
+
+  if (hasFunctionDeclarator(node)) {
+    return { skipChildren: true };
+  }
+
+  if (isInsideClassLike(node)) {
+    return;
+  }
+
+  addCppDeclarationSymbols(node, filePath, symbols, isCppConstantDeclaration(node) ? 'constant' : 'global');
+  return { skipChildren: true };
+}
+
+function addCppDeclarationSymbols(
+  node: Parser.SyntaxNode,
+  filePath: string,
+  symbols: IAnalysisSymbol[],
+  kind: 'constant' | 'global' | 'local',
+): void {
+  for (const nameNode of getDeclaratorNameNodes(node)) {
+    addNamedSymbol(symbols, filePath, kind, nameNode, node);
+  }
+}
+
+function isCppConstantDeclaration(node: Parser.SyntaxNode): boolean {
+  return /\b(?:const|constexpr)\b/.test(node.text.split(/[=;]/, 1)[0] ?? '');
+}
