+++ packages/core/src/treeSitter/runtime/analyzeCpp/symbolDeclaratorNames.ts
@@ -1,0 +1,2 @@
+export { getDeclaratorNameNodes } from './symbolDeclaratorCandidates';
+export { getDeclaratorNameNode } from './symbolDeclaratorNameNode';
